[Refractor] Slf4j로 로거 수정, 테스트명 수정
build.gradle에 Slf4j 관련 의존성 추가, 혼동되는 테스트명 수정
