+++ BE/robocar/build.gradle
@@ -41,9 +41,9 @@ dependencies {
 	runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
 	runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5' // Jackson을 사용하여 JWT를 처리하기 위함
 
-	//로거
-	implementation 'ch.qos.logback:logback-classic:1.4.14'
-	implementation 'org.slf4j:slf4j-api:2.0.3'
+	//Slf4j
+	testCompileOnly 'org.projectlombok:lombok'
+	testAnnotationProcessor 'org.projectlombok:lombok'
 }
 
 tasks.named('test') {
